requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pandas
mysql-connector-python